Add CP/M output file format (.COM)
CP/M .COM files require a JMP ENTRY if the program entry point is
not the first byte of the output file. In order to handle this, ld80
creates an "empty" JMP before loading/linking the object files, and then
(after determining the entry point) decides whether the JMP ENTRY is
needed and deletes it if not.  If the JMP ENTRY is retained, the target
address is set after the program relocations, just prior to writing
the output file.